submission/source/requirements.txt
torch>=2.2.0
numpy>=1.24.0
PyYAML>=6.0
h5py>=3.10.0
tqdm>=4.66.0
